# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management System documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/stocks": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "List of all stocks",
                "operationId": "166396c784da14d7aa25eb5cf051d47d",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "product_id": {
                                                        "type": "integer"
                                                    },
                                                    "quantity": {
                                                        "type": "integer"
                                                    },
                                                    "type": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "id": 1,
                                            "product_id": 12,
                                            "quantity": 91,
                                            "type": "In",
                                            "created_at": "2024-06-26T07:51:10.000000Z"
                                        },
                                        {
                                            "id": 2,
                                            "product_id": 44,
                                            "quantity": 49,
                                            "type": "Out",
                                            "created_at": "2024-06-26T07:51:10.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
        "/api/v1/stocks/{stock}": {
            "get": {
                "tags": [
                    "Stock"
                ],
                "summary": "Get single stock",
                "operationId": "552785b15edfd604a857f063c4ef28f5",
                "parameters": [
                    {
                        "name": "stock",
                        "in": "path",
                        "description": "Stock id",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "product_id": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": "1"
                                                        },
                                                        "sku": {
                                                            "type": "string",
                                                            "example": "lorem-ipsum"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "name"
                                                        },
                                                        "price": {
                                                            "type": "integer",
                                                            "example": "10.20"
                                                        },
                                                        "stock": {
                                                            "type": "integer",
                                                            "example": 20
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "Voluptatum sequi odio sint dolorem consectetur nihil quasi."
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "example": "2024-06-26T07:51:10.000000Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "quantity": {
                                                    "type": "string",
                                                    "example": 41
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "In"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-06-26T07:51:10.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
        "/api/v1/stocks/in/{product}": {
            "post": {
                "tags": [
                    "Stock"
                ],
                "summary": "Stock In for product.",
                "operationId": "1596567f6bd925564c9cf846d014b717",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "quantity": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            },
                            "example": {
                                "quantity": "100"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "product": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": "1"
                                                        },
                                                        "sku": {
                                                            "type": "string",
                                                            "example": "lorem-ipsum"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "name"
                                                        },
                                                        "price": {
                                                            "type": "integer",
                                                            "example": "10.20"
                                                        },
                                                        "stock": {
                                                            "type": "integer",
                                                            "example": 20
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "Voluptatum sequi odio sint dolorem consectetur nihil quasi."
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "example": "2024-06-26T07:51:10.000000Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "quantity": {
                                                    "type": "string",
                                                    "example": "100"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "Out"
                                                },
                                                "created_at": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
        "/api/v1/stocks/out/{product}": {
            "post": {
                "tags": [
                    "Stock"
                ],
                "summary": "Stock Out for product.",
                "operationId": "34c9e34589c81039526080b66e849a05",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "quantity": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            },
                            "example": {
                                "quantity": "100"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "product": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": "1"
                                                        },
                                                        "sku": {
                                                            "type": "string",
                                                            "example": "lorem-ipsum"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "name"
                                                        },
                                                        "price": {
                                                            "type": "integer",
                                                            "example": "10.20"
                                                        },
                                                        "stock": {
                                                            "type": "integer",
                                                            "example": 20
                                                        },
                                                        "description": {
                                                            "type": "string",
                                                            "example": "Voluptatum sequi odio sint dolorem consectetur nihil quasi."
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "example": "2024-06-26T07:51:10.000000Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "quantity": {
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "Out"
                                                },
                                                "created_at": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": ""
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}